refactor(api): fix typo in filename
